requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
streamlit>=1.28.0
plotly>=5.17.0
Pillow>=9.0.0
numpy>=1.24.0,<2.0.0
pandas>=1.5.0
captum>=0.6.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0
